pyproject.toml
[tool.poetry]
name = "breastcancer-fuzzyinterface"
version = "0.1.0"
description = ""
authors = ["Tymek <tymekzapala12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
numpy1 = "^0.0.1"
matplotlib = "^3.8.4"
sympy = "^1.12"
scipy = "^1.13.0"
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
